Improve interpreter startup with IPC pipe
The Electron main process will now open a pipe
to the interpreter process, which in turn
will send the string "ready" over the pipe
when it is ready to evaluate expressions.

Previously, there was an arbitrary waiting time
of 2.4sec in the hope that the interpreter would
be ready by then.